Refactored routes, introduced a partial static class to ensure single responsibility. Added scopes to the jwt.
