+++ src/Authentica.Service.Identity.Tests/UnitTests/JsonWebTokenProviderTests.cs
@@ -16,11 +16,11 @@ public class JsonWebTokenProviderTests
     {
         // Arrange
         var mockProvider = new JsonWebTokenProviderMock();
-        mockProvider.Setup(p => p.TryCreateTokenAsync(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(), It.IsAny<int>(), It.IsAny<string>(), It.IsAny<IList<string>>()))
+        mockProvider.Setup(p => p.TryCreateTokenAsync(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(), It.IsAny<int>(), It.IsAny<string>(), It.IsAny<IList<string>>(), It.IsAny<IList<string>>()))
                     .ReturnsAsync(new JwtResult { Success = true, Token = "mockToken", Error = null });
 
         // Act
-        var result = await mockProvider.Object.TryCreateTokenAsync(_email, _jwtSecret, _issuer, _audience, _expires, _subject, _roles);
+        var result = await mockProvider.Object.TryCreateTokenAsync(_email, _jwtSecret, _issuer, _audience, _expires, _subject, _roles, null!);
 
         Assert.Multiple(() =>
         {
